git-svn-id: https://svn.apache.org/repos/asf/maven/surefire/tags/surefire-2.4@606990 13f79535-47bb-0310-9956-ffa450edef68

+++ surefire-booter/src/main/java/org/apache/maven/surefire/booter/ForkConfiguration.java
@@ -217,11 +217,13 @@ public class ForkConfiguration
      * @return
      * @throws IOException
      */
-    private static File createJar( List classPath )
+    private File createJar( List classPath )
         throws IOException
     {
         File file = File.createTempFile( "surefirebooter", ".jar" );
-        file.deleteOnExit();
+        if ( !debug ) {
+            file.deleteOnExit();
+        }
         ManifestJarWriter writer = new ManifestJarWriter(file);
 
         // we can't use StringUtils.join here since we need to add a '/' to
